services/backend/requirements.txt
fastapi==0.89.1
pymongo==4.3.3
uvicorn ==0.20.0
pandas==1.5.2
python-multipart == 0.0.5
pydantic == 1.10.4
sqlalchemy==1.4.22
alembic==1.9.1
tenacity == 8.1.0
greenlet == 2.0.1
pymongo==4.3.3
pymysql==1.0.2
python-jose==3.3.0
ecdsa==0.18.0
requests==2.28.1
urllib3==1.26.13
nltk==3.8.1
psycopg2-binary==2.9.9
tqdm==4.66.2
langdetect==1.0.9
langcodes==3.3.0
language_data==1.1
redis>=5.0.2
python-dotenv>=1.0.1
numpy>=1.26.4
schedule>=1.2.1
pyspellchecker>=0.8.1
autocorrect>=2.6.1
textblob==0.18.0.post0
asyncpg>=0.29.0
